+++ src/material/datepicker/testing/BUILD.bazel
@@ -21,13 +21,8 @@ filegroup(
 )
 
 ng_test_library(
-    name = "harness_tests_lib",
-    srcs = [
-        "calendar-harness-shared.spec.ts",
-        "date-range-input-harness-shared.spec.ts",
-        "datepicker-input-harness-shared.spec.ts",
-        "datepicker-toggle-harness-shared.spec.ts",
-    ],
+    name = "unit_tests_lib",
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/testing",
@@ -36,24 +31,6 @@ ng_test_library(
         "//src/material/datepicker",
         "@npm//@angular/forms",
         "@npm//@angular/platform-browser",
-    ],
-)
-
-ng_test_library(
-    name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = [
-            "date-range-input-harness-shared.spec.ts",
-            "datepicker-input-harness-shared.spec.ts",
-            "datepicker-toggle-harness-shared.spec.ts",
-            "calendar-harness-shared.spec.ts",
-        ],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
-        "//src/material/datepicker",
     ],
 )
